Move applyTextEdit to core
