+++ packages/foam-cli/src/commands/janitor.ts
@@ -5,8 +5,8 @@ import {
   generateLinkReferences,
   generateHeading,
   getKebabCaseFileName,
+  applyTextEdit
 } from 'foam-core';
-import { applyTextEdit } from '../utils/apply-text-edit';
 import { writeFileToDisk } from '../utils/write-file-to-disk';
 import { isValidDirectory } from '../utils';
